requirements.txt
fastAPI
gunicorn
uvicorn
Jinja2
pytest
requests
pre-commit
python-dotenv
python-multipart
aiofiles
pillow
pytesseract
chardet
